percolate-reading/pyproject.toml
[project]
name = "percolate-reading"
version = "0.1.0"
description = "Heavy multimedia processing services for Percolate - shared reading node"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "TBD" }
authors = [
    { name = "Percolation Labs" }
]
keywords = ["document-parsing", "embeddings", "ocr", "transcription", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.18",
    "websockets>=13.0",  # WebSocket support

    # Data validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # Database (for job management)
    # "percolate-rocks>=0.1.0",  # REM database with RocksDB (TODO: add when published)

    # Document parsing
    "kreuzberg>=0.1.0",  # Fast PDF/DOCX parsing (primary)
    "pypdf>=5.1.0",  # Fallback PDF parser
    "pdfplumber>=0.11.0",  # Table extraction fallback
    "openpyxl>=3.1.0",  # Excel parsing
    "python-docx>=1.1.0",  # DOCX parsing

    # OCR
    "pytesseract>=0.3.10",
    "pillow>=11.0.0",

    # Audio transcription
    "openai>=1.51.0",  # Whisper API (primary)
    "faster-whisper>=1.1.0",  # Local Whisper (fallback)
    "pydub>=0.25.1",  # Audio manipulation

    # Embeddings
    "fastembed>=0.4.2",  # Fast local embeddings
    "sentence-transformers>=3.3.0",  # Alternative embeddings

    # CLI
    "typer>=0.15.0",
    "rich>=13.9.0",

    # Observability
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "loguru>=0.7.3",

    # Utilities
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

gpu = [
    "torch[cuda]>=2.5.0",
]

[project.scripts]
percolate-reading = "percolate_reading.cli.main:app"

[project.urls]
Homepage = "https://github.com/percolation-labs/percolate"
Repository = "https://github.com/percolation-labs/percolate"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/percolate_reading"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]
